UI changes, context menus & conflicts feature
You can now check for conflicts on the main tagger menu, add them on tag creation or change them on tag review.
You can now right-click on the tags on the main list to edit them (if they are registered), create them (if unregistered) or just remove it (double-click also removes it).
Version of the app now shows on the preferences screen at the top right